Moved predicate to framework.
